requirements.txt
setuptools==69.0.2
# requirements_fixed.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core>=2.14.0  # Versión más reciente con mejor soporte
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

librosa==0.10.1
soundfile==0.12.1
numpy==1.26.2
scipy==1.11.4
praat-parselmouth==0.4.3  # Para F0, Jitter, Shimmer
pydub==0.25.1
resampy==0.4.2

# HTTP para descargar de S3
requests==2.31.0

noisereduce==3.0.3
fastdtw==0.3.4
ffmpeg-python==8.0_2 